+++ ui/client/reducers/selectors/other.ts
@@ -1,9 +1,10 @@
 import {createSelector} from "reselect"
-import {getProcessCategory, isArchived} from "./graph"
+import {getFetchedProcessDetails, getProcessCategory, isArchived} from "./graph"
 import {getLoggedUser} from "./settings"
 import {RootState} from "../index";
 import {NotificationsState} from "react-notification-system-redux";
 import {BackendNotificationState} from "../notifications";
+import {getProcessState, isProcessStateLoaded} from "./scenarioState";
 
 export interface Capabilities {
     write?: boolean,
@@ -24,3 +25,9 @@ export const getCapabilities = createSelector(
 export const getNotifications = (state: RootState): NotificationsState => state.notifications
 export const getBackendNotifications = (state: RootState): BackendNotificationState => state.backendNotifications
 
+export const getFetchedProcessState = createSelector(
+  getFetchedProcessDetails,
+  isProcessStateLoaded,
+  getProcessState,
+  (fetchedProcessDetails, isStateLoaded, processState) => isStateLoaded ? processState : fetchedProcessDetails?.state,
+)