+++ utils/components-testkit/src/main/scala/pl/touk/nussknacker/engine/util/test/TestScenarioRunner.scala
@@ -1,24 +1,55 @@
 package pl.touk.nussknacker.engine.util.test
 
 import cats.data.ValidatedNel
+import pl.touk.nussknacker.engine.api.component.ComponentDefinition
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError
 import pl.touk.nussknacker.engine.api.exception.NuExceptionInfo
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
+import pl.touk.nussknacker.engine.testmode.TestComponentsHolder
 import pl.touk.nussknacker.engine.util.test.TestScenarioRunner.RunnerResult
 
 import scala.reflect.ClassTag
 
+/**
+  * This is entrypoint for all available test runners. Use:
+  *
+  * import SomeSpecificTestRunner._
+  * TestScenarioRunner.specific()
+  *
+  * to run tests. For example for kafka-lite it will be:
+  *
+  * import LiteKafkaTestScenarioRunner._
+  * TestScenarioRunner.kafkaLiteBased()
+  *
+  */
 object TestScenarioRunner {
+
   type RunnerResult[R] = ValidatedNel[ProcessCompilationError, RunResult[R]]
+
+  // Maybe we should replace ids with more meaningful: test-data, rest-result?
+  val testDataSource = "source"
+  val noopSource = "noopSource"
+  // Maybe we should unify those two test result nodes?
+  val testResultSink = "sink"
+  def testResultService: String = TestComponentsHolder.invocationCollectorComponentName
+
 }
 
 /**
-  * This is *experimental* API, currently it allows only for simple use case - synchronous invocation of test data
+  * This is *experimental* API
   *
   * This is common trait for all various engines - where each engine can have different entry method provided by convention runWith***Data.
   * For example see implementation on LiteKafkaTestScenarioRunner and LiteTestScenarioRunner
   */
 trait TestScenarioRunner
+
+trait TestScenarioRunnerBuilder[R <: TestScenarioRunner, B <: TestScenarioRunnerBuilder[R, _]] {
+
+  def withExtraComponents(components: List[ComponentDefinition]): B
+
+  def build(): R
+
+}
 
 trait ClassBasedTestScenarioRunner extends TestScenarioRunner {
   //todo add generate test data support